move SqliteVecQueueDatabase to db directory
